Fix parameters type of subimage_view functions in core and dynamic_image
Align with `subsampled_view` overloads accepting `point_t` and `coord_t`
Both types are based on `std::ptrdiff_t`. Fixes compilation warnings.
Tidy up `subimage_view` and `subsampled_view` with trailing return.
Add tests for `subimage_view` in core and dynamic_image extension.
(First tests for the dynamic_image, hurray! :))

Add explicit cast to double in division operator for point<T> to
fix compilation warnings.

+++ include/boost/gil/point.hpp
@@ -177,8 +177,8 @@ auto operator/(point<T> const& p, D d)
     using result_type = typename detail::std_common_type<T, D>::type;
     if (d < 0 || 0 < d)
     {
-        double const x = p.x / static_cast<double>(d);
-        double const y = p.y / static_cast<double>(d);
+        double const x = static_cast<double>(p.x) / static_cast<double>(d);
+        double const y = static_cast<double>(p.y) / static_cast<double>(d);
         return point<result_type>{
             static_cast<result_type>(iround(x)),
             static_cast<result_type>(iround(y))};
